Numerous performance changes

+++ catscripts/Entertainment/Cartoons/Dora_the_Explorer/Dora.sh
@@ -1,14 +1,24 @@
 #!/bin/bash
 
-egrep -i 'Dora(| )the(| )Explorer' newpages.txt >> Dora.txt
+KEYWORDS_DORA="Dora(| )the(| )Explorer"
 
-DORA=`stat --print=%s Dora.txt`
-
-if [ $DORA -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Dora.txt"
-  export CATNAME="Dora the Explorer"
-  $CATEGORIZE
-fi
 
-rm Dora.txt
+  debug_start "Dora the Explorer"
+
+  DORA="$(egrep -i "$KEYWORDS_DORA" newpages.txt)"
+
+  if [ "$DORA" != "" ];
+  then
+    printf "%s" "$DORA" > Dora.txt
+    export CATFILE="Dora.txt"
+    export CATNAME="Dora the Explorer"
+    $CATEGORIZE
+    rm Dora.txt
+    unset DORA
+  fi
+
+  debug_end "Dora the Explorer"
+
+fi
