+++ catscripts/Entertainment/Video_games/Video_game_consoles/Microsoft/Xbox.sh
@@ -3,15 +3,23 @@
 KEYWORDS_XBOX="X(|-)box"
 KEYWORDS_XBOX_EXCLUDE="X(|-)box(| )(1\b|360|One|Live)"
 
-egrep -i "$KEYWORDS_XBOX" newpages.txt | egrep -iv "$KEYWORDS_XBOX_EXCLUDE" >> Xbox.txt
-
-XBOX=`stat --print=%s Xbox.txt`
-
-if [ $XBOX -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Xbox.txt"
-  export CATNAME="Xbox"
-  $CATEGORIZE
-fi
 
-rm Xbox.txt
+  debug_start "Xbox"
+
+  XBOX="$(egrep -i "$KEYWORDS_XBOX" newpages.txt | egrep -iv "$KEYWORDS_XBOX_EXCLUDE")"
+
+  if [ "$XBOX" != "" ];
+  then
+    printf "%s" "$XBOX" > Xbox.txt
+    export CATFILE="Xbox.txt"
+    export CATNAME="Xbox"
+    $CATEGORIZE
+    rm Xbox.txt
+    unset XBOX
+  fi
+
+  debug_end "Xbox"
+
+fi
