+++ catscripts/Transportation/Transportation.sh
@@ -1,26 +1,36 @@
 #!/bin/bash
 
-egrep -i 'transportation' newpages.txt >> Transportation.txt
-egrep -i 'school bus' newpages.txt | egrep -iv 'Magic School Bus' >> Transportation.txt
+KEYWORDS_TRANSPORTATION="Transportation"
+KEYWORDS_SCHOOLBUS="School(| )Bus"
+KEYWORDS_SCHOOLBUS_EXCLUDE="Magic(| )School(| )Bus"
+KEYWORDS_TRANSPORTATION_EXCLUDE="$KEYWORDS_SCHOOLBUS"
 
-TRANSPORTATION=`stat --print=%s Transportation.txt`
+TRANSPORTATIONDIR="./catscripts/Transportation"
+#$TRANSPORTATIONDIR/Airplanes/Airplanes.sh
+$TRANSPORTATIONDIR/Cars/Cars.sh
+$TRANSPORTATIONDIR/SUVs/SUVs.sh
+#$TRANSPORTATIONDIR/Trucks/Trucks.sh
+#$TRANSPORTATIONDIR/Vans/Vans.sh 
+$TRANSPORTATIONDIR/Vehicle_manufacturers/Manufacturers.sh
 
-if [ $TRANSPORTATION -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Transportation.txt"
-  export CATNAME="Transportation"
-  $CATEGORIZE
+
+  debug_start "Transportation"
+
+  TRANSPORTATION="$(egrep -i "$KEYWORDS_TRANSPORTATION" newpages.txt | egrep -iv "$KEYWORDS_TRANSPORTATION_EXCLUDE")"
+  SCHOOLBUS="$(egrep -i "$KEYWORDS_SCHOOLBUS" newpages.txt | egrep -iv "$KEYWORDS_SCHOOLBUS_EXCLUDE")"
+
+  if [ "$TRANSPORTATION" != "" ];
+  then
+    printf "%s" "$TRANSPORTATION" > Transportation.txt
+    export CATFILE="Transportation.txt"
+    export CATNAME="Transportation"
+    $CATEGORIZE
+    rm Transportation.txt
+    unset TRANSPORTATION
+  fi
+
+  debug_end "Transportation"
+
 fi
-
-CURRENTDIR="./catscripts/Transportation"
-
-
-
-#$CURRENTDIR/Airplanes/Airplanes.sh
-$CURRENTDIR/Cars/Cars.sh
-$CURRENTDIR/SUVs/SUVs.sh
-#$CURRENTDIR/Trucks/Trucks.sh
-#$CURRENTDIR/Vans/Vans.sh 
-$CURRENTDIR/Vehicle_manufacturers/Manufacturers.sh
-
-rm Transportation.txt
